\name{TilePlot-package}
\alias{TilePlot-package}
\alias{TilePlot}
\alias{tileplot}
\docType{package}
\title{
Package for analysis of Functional Gene Tiling DNA Microarray}
\description{
This package is intended for processing the output from functional gene tiling DNA microarray experiments. It produces hybridization pattern plots for each gene on the array, and statistics for each gene including mean probe intensity, median probe intensity, bright probe fraction, bright segment length dependent score, bright probe mean intensity, and bright probe median intensity. Output is generated in order of bright segment length dependent score in both a latex/eps format and tab-delimited text file. The package works in three modes: single array, comparison of two arrays, and comparison of two arrays with loess normalization. 
}
\details{
\tabular{ll}{
Package: \tab TilePlot\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-09-24\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ian Marshall <ianpgm@stanford.edu>
}
\references{
http://www.stanford.edu/~ianpgm/tileplot/
}
